+++ src/flockwave/server/ext/mavlink/driver.py
@@ -1542,6 +1542,8 @@ class MAVLinkUAV(UAVBase):
         ):
             yield event
 
+        self._clear_autopilot_capabilities()
+
     def handle_message_autopilot_version(self, message: MAVLinkMessage):
         """Handles an incoming MAVLink AUTOPILOT_VERSION message targeted at
         this UAV.
@@ -2393,6 +2395,15 @@ class MAVLinkUAV(UAVBase):
 
         # Reset our internal state object of the compass calibration procedure
         self.compass_calibration.reset()
+
+    def _clear_autopilot_capabilities(self) -> None:
+        """Clears the cached autopilot capabilities and firmware version number
+        of the UAV.
+
+        This function should be called after a firmware update to ensure that
+        we query the new firmware version after the update.
+        """
+        self._last_messages.pop(MAVMessageType.AUTOPILOT_VERSION, None)
 
     async def _request_autopilot_capabilities(self) -> None:
         """Sends a request to the autopilot to send its capabilities via MAVLink