+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [main]
+
+### Fixed
+
+- MAVLink autopilot version number is now queried again after a firmware update
+  instead of returning the previously cached value.
+
 ## [2.28.0] - 2025-06-07
 
 ### Added